requirements.txt
anyio==4.6.2.post1
asttokens==2.4.1
attrs==24.2.0
cattrs==24.1.2
cellpose==3.0.11
certifi==2024.8.30
cffi==1.17.1
cgal==6.0.1.post202410241521
chardet==5.2.0
charset-normalizer==3.4.0
colorlog==6.9.0
comm==0.2.2
contourpy==1.3.0
cryptography==43.0.3
cycler==0.12.1
debugpy==1.8.5
decorator==5.1.1
deepdiff==8.0.1
dipy==1.9.0
distro==1.9.0
embreex==2.17.7.post5
executing==2.1.0
fastremap==1.15.0
filelock==3.16.0
fonttools==4.53.1
fsspec==2024.9.0
h11==0.14.0
h5py==3.12.1
hausdorff==0.2.6
httpcore==1.0.7
httpx==0.27.2
idna==3.10
imagecodecs==2024.6.1
imageio==2.35.1
intel-openmp==2023.1.0
ipykernel==6.29.5
ipython==8.27.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.2
jupyter_core==5.7.2
kiwisolver==1.4.7
lazy_loader==0.4
libigl==2.5.1
llvmlite==0.43.0
lxml==5.3.0
manifold3d==3.0.0
mapbox_earcut==1.0.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
meshio==5.3.5
meshplex==0.19.7
mkl==2023.1.0
mpmath==1.3.0
natsort==8.4.0
nest-asyncio==1.6.0
networkx==3.3
nibabel==5.3.2
npx==0.1.6
numba==0.60.0
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.68
nvidia-nvtx-cu12==12.1.105
opencv-contrib-python==4.5.5.64
opencv-python==4.8.0.74
opencv-python-headless==4.10.0.84
optimesh==0.10.4
orderly-set==5.2.2
packaging==24.1
parso==0.8.4
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.3.2
point-cloud-utils==0.31.0
pooch==1.8.2
POT==0.9.5
potpourri3d==1.1.0
prompt_toolkit==3.0.47
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-machineid==0.6.0
pyacvd==0.3.1
pycollada==0.8
pycparser==2.22
pygeodesic==0.1.9
Pygments==2.18.0
pykdtree==1.3.13
pypardiso==0.4.6
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-package-info==0.0.10
pyvista==0.44.1
pyzmq==26.2.0
referencing==0.35.1
requests==2.32.3
requests-cache==1.2.1
rich==13.9.4
rich-argparse==1.6.0
robust_laplacian==1.0.0
roifile==2024.5.24
rpds-py==0.21.0
Rtree==1.3.0
scikit-image==0.24.0
scikit-learn==1.5.2
scipy==1.14.1
scooby==0.10.0
setuptools==74.1.2
setuptools-scm==8.1.0
shapely==2.0.6
SimpleITK==2.4.0
six==1.16.0
sniffio==1.3.1
stack-data==0.6.3
stonefish-license-manager==0.4.37
svg.path==6.3
sympy==1.13.2
tbb==2021.13.1
threadpoolctl==3.5.0
tifffile==2024.8.30
torch==2.4.1
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
transforms3d==0.4.2
trimesh==4.5.2
triton==3.0.0
trx-python==0.3
typing_extensions==4.12.2
u-Unwrap3D @ file:///home/c.soubrier/Documents/UBC_Vancouver/Projets_recherche/Shape_analysis/unwrap/u-unwrap3D
url-normalize==1.4.3
urllib3==2.2.3
vhacdx==0.0.8.post1
vtk==9.3.1
wcwidth==0.2.13
x21==0.5.3
xatlas==0.0.9
xxhash==3.5.0
